api/urls.py
from django.urls import path, include
from rest_framework_simplejwt import views as jwt_views
from rest_framework.routers import DefaultRouter

from . import viewsets


router = DefaultRouter()
router.register('user', viewsets.UserViewSet)
router.register('post', viewsets.PostViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('token/', jwt_views.TokenObtainPairView.as_view(), name='token-obtain-pair'),
    path('token/refresh/', jwt_views.TokenRefreshView.as_view(), name='toke-refresh'),
]
