requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
python-dotenv
requests
